Patched a server-side crash
The crash occurs when shooting a grappling vine at a Nether portal, when there is a player on the other side. It seems like the cause is that the `userU` field isn't saved in NBT, causing it to become null after the game serializes and then deserializes the entity in order to add it to another dimension.
`userU` doesn't seem to be used on the client side (hurtTarget is only called on server), so it doesn't need to be synced anyway.